+++ packages/zod-dbs-cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # zod-dbs-cli
 
+## 2.1.0
+
+### Minor Changes
+
+- Add option to disable index.ts (barrel file) generation.
+
+### Patch Changes
+
+- Updated dependencies
+  - zod-dbs-core@2.1.0
+  - zod-dbs@2.1.0
+
 ## 2.0.7
 
 ### Patch Changes