+++ packages/zod-dbs-pg/CHANGELOG.md
@@ -1,5 +1,12 @@
 ## v1.0.6
 
+## 2.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - zod-dbs-core@2.1.0
+
 ## 2.0.4
 
 ### Patch Changes